// types/cloudcasts.d.ts
// Generated by https://quicktype.io

export interface CloudcastsPage {
    paging?: Paging;
    data: Datum[];
    name: string;
}

export interface Datum {
    tags: Tag[];
    play_count: number;
    user: User;
    key: string;
    created_time: string;
    audio_length: number;
    slug: string;
    favorite_count: number;
    listener_count: number;
    name: string;
    url: string;
    pictures: { [key: string]: string };
    repost_count: number;
    updated_time: string;
    comment_count: number;
}

export interface Tag {
    url: TagURL;
    name: TagName;
    key: TagKey;
}

export enum TagKey {
    DiscoverChill = "/discover/chill/",
    DiscoverChillOut = "/discover/chill-out/",
    DiscoverChillout = "/discover/chillout/",
    DiscoverDowntempo = "/discover/downtempo/",
    DiscoverLounge = "/discover/lounge/",
}

export enum TagName {
    Chill = "Chill",
    ChillOut = "Chill Out",
    Chillout = "Chillout",
    Downtempo = "Downtempo",
    Lounge = "Lounge",
}

export enum TagURL {
    HTTPSWWWMixcloudCOMDiscoverChill = "https://www.mixcloud.com/discover/chill/",
    HTTPSWWWMixcloudCOMDiscoverChillOut = "https://www.mixcloud.com/discover/chill-out/",
    HTTPSWWWMixcloudCOMDiscoverChillout = "https://www.mixcloud.com/discover/chillout/",
    HTTPSWWWMixcloudCOMDiscoverDowntempo = "https://www.mixcloud.com/discover/downtempo/",
    HTTPSWWWMixcloudCOMDiscoverLounge = "https://www.mixcloud.com/discover/lounge/",
}

export interface User {
    url: UserURL;
    username: Username;
    name: UserName;
    key: UserKey;
    pictures: { [key: string]: string };
}

export enum UserKey {
    LucForlorn = "/Luc_Forlorn/",
}

export enum UserName {
    LucForlorn = "Luc Forlorn",
}

export enum UserURL {
    HTTPSWWWMixcloudCOMLucForlorn = "https://www.mixcloud.com/Luc_Forlorn/",
}

export enum Username {
    LucForlorn = "Luc_Forlorn",
}

export interface Paging {
    previous: string;
    next: string;
}